# doc/api.yaml
openapi: 3.0.3 
info:
  title: WASAText
  description: |-
    This OpenAPI document describes how WASAText works.
  version: "1"
  contact:
    name: Tommaso Salvi
    email: salvi.1960045@studenti.uniroma1.it
...
servers: 
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
  schemas:
...
    Error500:
      type: object
      description: "Schema for a server error (500)."
      properties:
        error:
          type: string
          description: "The code of the error"
          pattern: "^[a-zA-Z0-9 .,!?'-]+$"
          minLength: 3
          maxLength: 1000
...
security:
  - bearerAuth: []
paths: 
...
  /groups/{Group_id}/users/{User_id}:
    post:
      tags: ["groups"]
      summary: Add a user in a group
      description: Adds a user in a specific group.
      operationId: addToGroup
      parameters:
        - name: Group_id
          description: The integer that describes the Group_id.
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            description: "The ID of the group."
        - name: User_id
          description: The integer that describes the User_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            description: "The ID of the user."
      requestBody:
        description: Details of the user to add 
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body for managing group users.
              required:
                - User_id
                - action
              properties:
                User_id:
                  type: integer
                  description: "ID of the user to add"
                  minimum: 1
                  maximum: 1000
      responses:
        '201':
          description: User successfully added in the group.
          content:
            application/json:
              schema:
                type: object
                description: Schema for a successful response.
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    description: "Indicates the operation was successful."
                  message:
                    type: string
                    description: "A confirmation message."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '400':
          description: Invalid data (Group_id, User_id).
          content:
            application/json:
              schema:
                type: object
                description: Schema for invalid data.
                properties:
                  error:
                    type: string
                    description: "Details of the error."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '404':
          description: Group or user not found.
          content:
            application/json:
              schema:
                type: object
                description: Schema for error caused by not found.
                properties:
                  error:
                    type: string
                    description: "Group or user not found."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'

    delete:
      tags: ["groups"]
      summary: Remove a user by a group
      description: Removes a user in a specific group.
      operationId: leaveGroup
      parameters:
        - name: Group_id
          description: The integer that describes the Group_id.
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            description: "The ID of the group."
        - name: User_id
          description: The integer that describes the User_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            description: "The ID of the user."
      responses:
        '201':
          description: User successfully removed by the group.
          content:
            application/json:
              schema:
                type: object
                description: Schema for a successful response.
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    description: "Indicates the operation was successful."
                    example: true
                  message:
                    type: string
                    description: "A confirmation message."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '400':
          description: Invalid data (Group_id, User_id, or action).
          content:
            application/json:
              schema:
                type: object
                description: Schema for invalid data.
                properties:
                  error:
                    type: string
                    description: "Details of the error."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '404':
          description: Group or user not found.
          content:
            application/json:
              schema:
                type: object
                description: Schema for error caused by not found.
                properties:
                  error:
                    type: string
                    description: "Group or user not found."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  /groups/{Group_id}:
    parameters:
        - name: Group_id
          in: path
          description: Integer that represents Group_id
          required: true
          schema:
            type: integer
            description: "The unique ID of the group."
            minimum: 1
            maximum: 1000
    patch:
      tags: ["groups"]
      summary: Update group name
      description: Allows updating certain group fields (like the group name).
      operationId: setGroupName
      requestBody:
        description: The new name for the group.
        required: true
        content:
          application/json-patch+json:
            schema:
              type: object
              required:
                - groupName
              description: "Schema for renaming a group"
              properties:
                groupName:
                  type: string
                  description: "New group name."
                  pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                  minLength: 3
                  maxLength: 1000
      responses:
        '200':
          description: Group name updated successfully.
          content:
            application/json:
              schema:
                type: object
                description: Schema for positive name update.
                required:
                  - Group_id
                  - groupName
                properties:
                  Group_id:
                    type: integer
                    description: "The unique ID of the group."
                    minimum: 1
                    maximum: 1000
                  groupName:
                    type: string
                    description: "The updated name of the group."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '400':
          description: Invalid group name or Group_id provided.
          content:
            application/json:
              schema:
                type: object
                description: Schema for negative response caused by invalid Group_id
                properties:
                  error:
                    type: string
                    description: "Error details."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '404':
          description: Group not found.
          content:
            application/json:
              schema:
                type: object
                description: Schema for error caused by not found.
                properties:
                  error:
                    type: string
                    description: "The specified group does not exist."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '500':
          description: Internal server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'

    delete:
      tags: ["groups"]
      summary: Delete a group
      description: Allows the deletion of a group.
      operationId: deleteGroup
      responses:
        '200':
          description: Group deleted successfully.
          content:
            application/json:
              schema:
                type: object
                description: Schema for successful deletion response.
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the deletion was successful."
                    example: true
                  message:
                    type: string
                    description: "A confirmation message."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '403':
          description: User not authorized to delete the group.
          content:
            application/json:
              schema:
                type: object
                description: Schema for unauthorized response.
                properties:
                  error:
                    type: string
                    description: "Error details."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '404':
          description: Group not found.
          content:
            application/json:
              schema:
                type: object
                description: Schema for group not found response.
                properties:
                  error:
                    type: string
                    description: "The specified group does not exist."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '500':
          description: Internal server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  /groups/{Group_id}/photo: 
    parameters:
        - name: Group_id
          in: path
          required: true
          description: "The unique ID of the group."
          schema:
            type: integer
            minimum: 1
            maximum: 1000
    put:
      tags: ["groups"]
      summary: Add or update a group photo
      description: Allows a user to upload or update a group's profile photo.
      operationId: setGroupPhoto
      requestBody:
        description: The group photo to upload.
        required: true
        content:
          multipart/form-data:
            schema:
              minLength: 1
              maxLength: 1000
              type: object
              required:
                - photo
              description: Schema for uploading a photo.
              properties:
                photo:
                  type: string
                  format: binary
                  description: "The image file for the group's profile photo."
      responses:
        '200':
          description: Group photo uploaded successfully.
          content:
            application/json:
              schema:
                type: object
                description: Schema for successful upload response.
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the upload was successful."
                    example: true
                  message:
                    type: string
                    description: "A confirmation message."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '400':
          description: Invalid file upload or bad request.
          content:
            application/json:
              schema:
                type: object
                description: Schema for invalid upload response.
                properties:
                  error:
                    type: string
                    description: "Error details."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '403':
          description: User not authorized to upload a photo.
          content:
            application/json:
              schema:
                type: object
                description: Schema for unauthorized response.
                properties:
                  error:
                    type: string
                    description: "Error details."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '404':
          description: Group not found.
          content:
            application/json:
              schema:
                type: object
                description: Schema for group not found response.
                properties:
                  error:
                    type: string
                    description: "The specified group does not exist."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '500':
          description: Internal server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'

    get:
      tags: ["groups"]
      summary: Retrieve a group's profile photo
      description: Fetches the profile photo of a specified group.
      operationId: getGroupPhoto
      responses:
        '200':
          description: Group photo returned successfully.
          content:
            application/json:
              schema:
                type: object
                description: Schema for successful returned response.
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the return was successful."
                    example: true
                  message:
                    type: string
                    description: "A confirmation message."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '400':
          description: Invalid group ID format.
          content:
            application/json:
              schema:
                type: object
                description: Schema for invalid request response.
                properties:
                  error:
                    type: string
                    description: "Error details."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '404':
          description: Group photo not found.
          content:
            application/json:
              schema:
                type: object
                description: Schema for group photo not found response.
                properties:
                  error:
                    type: string
                    description: "The group does not have a profile photo."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '500':
          description: Internal server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...